requirements.txt
influxdb_client
python-dotenv
pytest
influxdb3-python
pandas